Move direction to upper package.
