+++ TelepresenceBot/mobile/src/main/java/com/novoda/tpbot/human/ControllerListener.java
@@ -1,6 +1,7 @@
 package com.novoda.tpbot.human;
 
 import com.novoda.notils.logger.simple.Log;
+import com.novoda.tpbot.Direction;
 
 interface ControllerListener {
     ControllerListener NO_OP = new ControllerListener() {
